backend/ai/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
scikit-learn==1.5.2
numpy==2.1.1
requests==2.32.3
openai==1.47.0
google-generativeai==0.7.2
transformers==4.44.2
torch==2.4.1